+++ src/main/java/com/dolgikh/scriptorium/util/exceptions/notfoundexceptions/GenreNotFoundException.java
@@ -1,7 +1,7 @@
 package com.dolgikh.scriptorium.util.exceptions.notfoundexceptions;
 
 public class GenreNotFoundException extends ResourceNotFoundException {
-    public GenreNotFoundException(int id) {
+    public GenreNotFoundException(long id) {
         super("Genre", id);
     }
 }
